Enum figures::Rect

pub enum Rect<T, Unit> {
    Sized(SizedRect<T, Unit>),
    Extents(ExtentsRect<T, Unit>),
}

A 2d rectangle. This type may internally be represented with a SizedRect or an ExtentsRect. All rect types implement Rectlike.

Variants

Sized(SizedRect<T, Unit>)

A SizedRect.

Extents(ExtentsRect<T, Unit>)

An ExtentsRect.

Implementations

impl<T, Unit> Rect<T, Unit>

pub fn sized(origin: Point<T, Unit>, size: Size<T, Unit>) -> Self

Returns a new rect using origin and size.

pub fn extents(origin: Point<T, Unit>, extent: Point<T, Unit>) -> Self

Returns a new rect using points origin and extent.

impl<T, Unit> Rect<T, Unit> where
T: NumCast + Copy,

pub fn try_cast<NewT: NumCast + Copy>(&self) -> Option<Rect<NewT, Unit>>

Attempts to cast T to NewT. If unsuccessful, None is returned.

pub fn cast<NewT: NumCast + Copy>(&self) -> Rect<NewT, Unit>

Casts T to NewT.

Panics

Panics if casting fails.

impl<T, Unit> Rect<T, Unit> where
T: Copy,

pub fn cast_unit<NewUnit>(&self) -> Rect<T, NewUnit>

Returns this value with the new unit. Does not affect the underlying value.

impl<T, Unit> Rect<T, Unit> where
T: Ceil + Floor,

pub fn round_out(self) -> Self

Returns a new rectangle that rounds the origin down using floor and rounds the extent/size out using ceil.

pub fn round_in(self) -> Self

Returns a new rectangle that rounds the origin up using ceil and rounds the extent/size in using floor.