Struct figures::Fraction

#[repr(C)]
pub struct Fraction { /* private fields */ }

A fraction type for predictable integer-based math.

Internally this type uses 32 bits of data to represent a fraction:

1 bit of data for the positive/negative sign.

15 bits of data for the numerator

16 bits of data for the denominator

Many math operations are performed using temporary 32-bit values for the fraction, simplifing at the end of the operation. This prevents overflows, but does not prevent precision loss. We can see this by purposely buliding fractions that are hard to represent:

use figures::fraction;

assert_eq!(
    fraction!(1 / 32719) + fraction!(1 / 32749),
    fraction!(2 / 32749)
);

The above example adds fractions that have denominators using two largest primes that fit in 16 bits. The result should be 65,468/1,071,514,531, but that denominator is clearly too large. Because the fractions being added had prime numbers for their denominators, there is no way to reduce this fraction without losing information. For this particular example, a precision loss of ~2.8e-8 occurs.

However, in 2d graphics programming, it’s rare to be working with irrational numbers outside of angles represented in radians.

Implementations

impl Fraction

pub const MAX: Self = _

The maximum value representable by this type.

pub const MIN: Self = _

The minimum value representable by this type.

pub const ONE: Self = _

A fraction equivalent to 1.

pub const PI: Self = _

A fractional approximation of Pi, accurate to within 2.67e-7.

pub const ZERO: Self = _

A fraction equivalent to 0.

pub const fn new_whole(whole_number: i16) -> Self

Returns a new fraction for a whole number.

pub fn new(numerator: i16, denominator: i16) -> Self

Returns a new fraction using the components provided.

denominator will be limited to the absolute value of i16::MIN.

pub const fn numerator(&self) -> i16

Returns the numerator of the fraction.

pub const fn denominator(&self) -> i16

Returns the denominator of the fraction.

pub const fn is_positive(&self) -> bool

Returns true if the fraction is positive (greater than zero).

Note: Zero is neither negative nor positive.

pub const fn is_zero(&self) -> bool

Returns true if the fraction is zero.

pub const fn is_negative(&self) -> bool

Returns true if the fraction is negative (less than zero).

Note: Zero is neither negative nor positive.

pub fn into_compound(self) -> (i16, Fraction)

Simplifies the fraction into a compound number.

use figures::Fraction;

assert_eq!(
    Fraction::new(1, 3).into_compound(),
    (0, Fraction::new(1, 3))
);
assert_eq!(
    Fraction::new(4, 3).into_compound(),
    (1, Fraction::new(1, 3))
);
assert_eq!(
    Fraction::new(-4, 3).into_compound(),
    (-1, Fraction::new(-1, 3))
);

Adding the number and the fraction back together will result in the original fraction.

use figures::Fraction;

let improper = Fraction::new(-4, 3);
let (whole, fraction) = improper.into_compound();
assert_eq!(Fraction::from(whole) + fraction, improper);

pub fn round(self) -> i16

Rounds this fraction to the nearest whole number.

pub fn round_with_amount(self) -> (i16, Fraction)

Rounds this fraction to the nearest whole number, returning the fractional component that was rounded away. This fraction can be added to the whole number to reconstruct the original fraction.

use figures::Fraction;

let (whole, fraction) = Fraction::new(5, 3).round_with_amount();
assert_eq!(whole, 2);
assert_eq!(fraction, Fraction::new(-1, 3));
assert_eq!(Fraction::new_whole(whole) + fraction, Fraction::new(5, 3));

let (whole, fraction) = Fraction::new(-5, 3).round_with_amount();
assert_eq!(whole, -2);
assert_eq!(fraction, Fraction::new(1, 3));
assert_eq!(Fraction::new_whole(whole) + fraction, Fraction::new(-5, 3));

pub fn into_f32(self) -> f32

Returns this fraction as a floating point number.

pub const fn inverse(self) -> Self

Returns the inverse of this fraction.

pub const fn abs(self) -> Self

Returns the absolute value of this fraction.

pub fn atan(self) -> Angle

Returns the arctangent of this fraction.

This function is implemented using a lookup table and is an approximation.

pub fn atan2(self, other: Self) -> Angle

Returns the result of arctan(self/other) while correctly handling negative numbers.