Struct figures::Point

pub struct Point<Unit> {
    pub x: Unit,
    pub y: Unit,
}

A coordinate in a 2d space.

Fields
x: Unit

The x-axis component.

y: Unit

The y-axis component

Implementations

impl<Unit> Point<Unit>

pub const fn new(x: Unit, y: Unit) -> Self

Returns a new point with the provided x and y components.

pub fn squared(i: Unit) -> Selfwhere Unit: Clone,

Returns a new point with both x and y initialized with i.

pub fn cast<NewUnit>(self) -> Point<NewUnit>where Unit: Into<NewUnit>,

Converts the contents of this point to NewUnit using From.

pub fn try_cast<NewUnit>(self) -> Result<Point<NewUnit>, Unit::Error>where Unit: TryInto<NewUnit>,

Converts the contents of this point to NewUnit using TryFrom.

Errors

Returns <NewUnit as TryFrom>::Error when the inner type cannot be converted. For this crate’s types, this genenerally will be TryFromIntError.

pub fn map<NewUnit>(self, map: impl FnMut(Unit) -> NewUnit) -> Point<NewUnit>

Maps each component to map and returns a new value with the mapped components.

pub fn dot(self, other: Point<Unit>) -> Unitwhere Unit: Mul<Output = Unit> + Add<Output = Unit>,

Returns the dot product of self and other.

pub fn magnitude(self) -> Unitwhere Unit: Mul<Output = Unit> + Add<Output = Unit> + Roots + Copy,

Returns the magnitude of self, which is the absolute distance from 0,0.

pub fn rotate_around(self, origin: Point<Unit>, angle: Angle) -> Point<Unit>where Unit: Copy + Add<Output = Unit> + Sub<Output = Unit> + Mul<Fraction, Output = Unit>,

Returns self rotated around origin by angle.

pub fn rotate_by(self, angle: Angle) -> Point<Unit>where Unit: Zero + Copy + Add<Output = Unit> + Sub<Output = Unit> + Mul<Fraction, Output = Unit>,

Returns self rotated around Point::ZERO by angle.