Struct figures::Rect

pub struct Rect<Unit> {
    pub origin: Point<Unit>,
    pub size: Size<Unit>,
}

A 2d area expressed as an origin (Point) and a Size.

Fields
origin: Point<Unit>

The origin of the rectangle

size: Size<Unit>

The size of the rectangle.

Implementations

impl<Unit> Rect<Unit>

pub const fn new(origin: Point<Unit>, size: Size<Unit>) -> Self

Returns a new rectangle.

pub fn from_extents(p1: Point<Unit>, p2: Point<Unit>) -> Selfwhere Unit: Unit,

Returns a new rectangle using the given points to form the top-left and bottom-right of the rectangle.

The order of the parameters does not matter. The minimum values will form the top-left and the maximum values will form the bottom-right.

pub fn expand_rounded(self) -> Selfwhere Unit: Round + Unit,

Expands this rect to the nearest whole number.

This function will never return a smaller rectangle.

pub fn map<NewUnit>(self, map: impl FnMut(Unit) -> NewUnit) -> Rect<NewUnit>

Maps each component to map and returns a new value with the mapped components.

pub fn inset(self, amount: impl Into<Unit>) -> Selfwhere Unit: Add<Unit, Output = Unit> + AddAssign<Unit> + SubAssign<Unit> + Copy,

Returns a rectangle that has been inset by amount on all sides.

pub fn cast<NewUnit>(self) -> Rect<NewUnit>where NewUnit: From<Unit>,

Converts the contents of this point to NewUnit using From.

pub fn try_cast<NewUnit>(self) -> Result<Rect<NewUnit>, NewUnit::Error>where NewUnit: TryFrom<Unit>,

Converts the contents of this rect to NewUnit using TryFrom.

Errors

Returns <NewUnit as TryFrom>::Error when the inner type cannot be converted. For this crate’s types, this genenerally will be

pub fn contains(&self, point: Point<Unit>) -> boolwhere Unit: Unit,

Returns true if this rect contains point.

pub fn intersects(&self, other: &Self) -> boolwhere Unit: Add<Output = Unit> + Ord + Copy,

Returns true if the areas of self and other overlap.

This function does not return true if the edges touch but do not overlap.

use figures::{Point, Rect, Size};

let a: Rect<i32> = Rect::new(Point::new(1, 1), Size::new(2, 2));
let b = Rect::new(Point::new(2, 2), Size::new(1, 1));
assert!(a.intersects(&b));
let c = Rect::new(Point::new(3, 1), Size::new(1, 1));
assert!(!a.intersects(&c));

pub fn intersection(&self, other: &Self) -> Option<Rect<Unit>>where Unit: Unit,

Returns the overlapping rectangle of self and other. If the rectangles do not overlap, None will be returned.

use figures::{Point, Rect, Size};

let a: Rect<i32> = Rect::new(Point::new(1, 1), Size::new(3, 3));
let b = Rect::new(Point::new(2, 2), Size::new(3, 3));
assert_eq!(
    a.intersection(&b),
    Some(Rect::new(Point::new(2, 2), Size::new(2, 2)))
);
let c = Rect::new(Point::new(4, 1), Size::new(1, 1));
assert_eq!(a.intersection(&c), None);

pub fn extent(&self) -> Point<Unit>where Unit: Unit,

Returns the non-origin point.

impl<Unit> Rect<Unit>where Unit: Add<Output = Unit> + Ord + Copy,

pub fn extents(&self) -> (Point<Unit>, Point<Unit>)

Returns the top-left and bottom-right points of this rectangle.

The first point returned will always be the top-right point, even if the size of the rectangle is negative.