Struct figures::Size

pub struct Size<Unit> {
    pub width: Unit,
    pub height: Unit,
}

A width and a height measurement.

Fields
width: Unit

The width component

height: Unit

The height component

Implementations

impl<Unit> Size<Unit>

pub const fn new(width: Unit, height: Unit) -> Self

Returns a new size of the given width and height.

pub fn squared(dimension: Unit) -> Selfwhere Unit: Copy,

Returns a new size using dimension for both width and height.

pub fn area(&self) -> <Unit as Mul>::Outputwhere Unit: Mul + Copy,

Returns the area of the rectangle.

pub fn cast<NewUnit>(self) -> Size<NewUnit>where NewUnit: From<Unit>,

Converts the contents of this size to NewUnit using From.

pub fn map<NewUnit>(self, map: impl FnMut(Unit) -> NewUnit) -> Size<NewUnit>

Maps each component to map and returns a new value with the mapped components.

pub fn try_cast<NewUnit>(self) -> Result<Size<NewUnit>, NewUnit::Error>where NewUnit: TryFrom<Unit>,

Converts the contents of this size to NewUnit using TryFrom.

Errors

Returns <NewUnit as TryFrom>::Error when the inner type cannot be converted. For this crate’s types, this genenerally will be