Struct figures::SizedRect

pub struct SizedRect<T, Unit> {
    pub origin: Point<T, Unit>,
    pub size: Size<T, Unit>,
}

A rectangle that uses a Point and a Size for representation.

Fields
origin: Point<T, Unit>

The origin of the rectangle.

size: Size<T, Unit>

The size of the rectangle.

Implementations

impl<T, Unit> SizedRect<T, Unit>

pub const fn new(origin: Point<T, Unit>, size: Size<T, Unit>) -> Self

Returns a new rectangle using origin and size.

impl<T, Unit> SizedRect<T, Unit> where
T: NumCast + Copy,

pub fn try_cast<NewT: NumCast + Copy>(&self) -> Option<SizedRect<NewT, Unit>>

Attempts to cast T to NewT. If unsuccessful, None is returned.

pub fn cast<NewT: NumCast + Copy>(&self) -> SizedRect<NewT, Unit>

Casts T to NewT.

Panics

Panics if casting fails.

impl<T, Unit> SizedRect<T, Unit> where
T: Copy,

pub fn cast_unit<NewUnit>(&self) -> SizedRect<T, NewUnit>

Returns this value with the new unit. Does not affect the underlying value.

impl<T, Unit> SizedRect<T, Unit> where
T: Ceil + Floor,

pub fn round_out(self) -> Self

Returns a new rectangle that rounds the origin down using floor and rounds the size out using ceil.

pub fn round_in(self) -> Self

Returns a new rectangle that rounds the origin up using ceil and rounds the size in using floor.