Trait figures::FloatConversion

pub trait FloatConversion {
    type Float;

    // Required methods
    fn into_float(self) -> Self::Float;
    fn from_float(float: Self::Float) -> Self;
}

Converts a type to its floating point representation.

This trait exists because there is no trait in Rust to peform x as f32.

Required Associated Types

type Float

The type that represents this type in floating point form.

Required Methods

fn into_float(self) -> Self::Float

Returns this value in floating point form.

fn from_float(float: Self::Float) -> Self

Converts from floating point to this form.