Trait figures::IntoSigned

pub trait IntoSigned {
    type Signed;

    // Required method
    fn into_signed(self) -> Self::Signed;
}

Converts a value into its signed representation, clamping negative numbers to i32::MAX.

Required Associated Types

type Signed

The signed representation of this type.

Required Methods

fn into_signed(self) -> Self::Signed

Returns this value as an unsigned value. Values that are larger than can fit in an i32 are converted to i32::MAX.