Trait figures::Lp2D

pub trait Lp2D: FromComponents<Lp> {
    // Provided methods
    fn points(x: impl Into<FloatOrInt>, y: impl Into<FloatOrInt>) -> Self { ... }
    fn cm(x: impl Into<FloatOrInt>, y: impl Into<FloatOrInt>) -> Self { ... }
    fn mm(x: impl Into<FloatOrInt>, y: impl Into<FloatOrInt>) -> Self { ... }
    fn inches(x: impl Into<FloatOrInt>, y: impl Into<FloatOrInt>) -> Self { ... }
}

Constructors for types that are composed of two Lp components.

Provided Methods

fn points(x: impl Into<FloatOrInt>, y: impl Into<FloatOrInt>) -> Self

Returns a new value containing the x and y components converted into Lp using Lp::points/Lp::points_f.

fn cm(x: impl Into<FloatOrInt>, y: impl Into<FloatOrInt>) -> Self

Returns a new value containing the x and y components converted into Lp using Lp::cm/Lp::cm_f.

fn mm(x: impl Into<FloatOrInt>, y: impl Into<FloatOrInt>) -> Self

Returns a new value containing the x and y components converted into Lp using Lp::mm/Lp::mm_f.

fn inches(x: impl Into<FloatOrInt>, y: impl Into<FloatOrInt>) -> Self

Returns a new value containing the x and y components converted into Lp using Lp::inches/Lp::inches_f.

Object Safety

This trait is not object safe.