Trait figures::ScreenScale

pub trait ScreenScale {
    type Px;
    type UPx;
    type Lp;

    // Required methods
    fn into_px(self, scale: Fraction) -> Self::Px;
    fn from_px(px: Self::Px, scale: Fraction) -> Self;
    fn into_upx(self, scale: Fraction) -> Self::UPx;
    fn from_upx(px: Self::UPx, scale: Fraction) -> Self;
    fn into_lp(self, scale: Fraction) -> Self::Lp;
    fn from_lp(lp: Self::Lp, scale: Fraction) -> Self;
}

Converts this type into its measurement in Px and Lp.

Required Associated Types

type Px

This type when measuring with Px.

type UPx

This type when measuring with UPx.

type Lp

This type when measuring with Lp.

Required Methods

fn into_px(self, scale: Fraction) -> Self::Px

Converts this value from its current unit into device pixels (Px) using the provided scale factor.

fn from_px(px: Self::Px, scale: Fraction) -> Self

Converts from pixels into this type, using the provided scale factor.

fn into_upx(self, scale: Fraction) -> Self::UPx

Converts this value from its current unit into device pixels (UPx) using the provided scale factor.

fn from_upx(px: Self::UPx, scale: Fraction) -> Self

Converts from unsigned pixels into this type, using the provided scale factor.

fn into_lp(self, scale: Fraction) -> Self::Lp

Converts this value from its current unit into device independent pixels (Lp) using the provided scale factor.

fn from_lp(lp: Self::Lp, scale: Fraction) -> Self

Converts from Lp into this type, using the provided scale factor.